refactor: split the table
