+++ apps/admin/utils/date.ts
@@ -1,0 +1,10 @@
+export const formatRequestedAt = (iso: string) => {
+  const date = new Date(iso);
+  return date.toLocaleString("en-US", {
+    month: "short",
+    day: "numeric",
+    year: "numeric",
+    hour: "2-digit",
+    minute: "2-digit",
+  });
+};
